Initial mandelbrot.py
A recursive mandelbrot function is used to find out how many iterations it takes to prove the complex number is not part of the Mandelbrot set, or if it is. A black pixel is printed to the png if the number is approximated to be part of the Mandelbrot set, and a color corresponding to how close if not part of the set.
Variables are at the top of the program, are used to specify which part of the set, and with how much precision, it is going to be approximated.